# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HALO — Archangel Operations Layer API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List properties with computed money rollups
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PropertySummary"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Full property detail — brief, price list, jobs, money, paper, people
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Property detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PropertyDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateProperty
      tags: [properties]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PropertyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}/brief:
    post:
      operationId: writePropertyBrief
      tags: [properties]
      summary: AI-write the property brief from live data
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Brief written
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}/price-items:
    post:
      operationId: createPriceItem
      tags: [properties]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PriceItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PriceItem"
  /contacts:
    post:
      operationId: createContact
      tags: [properties]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
...
components:
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Property:
      type: object
      required: [id, name, status]
      properties:
        id: { type: string }
        name: { type: string }
        pmcName: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        units: { type: ["integer", "null"] }
        accessNotes: { type: ["string", "null"] }
        brief: { type: ["string", "null"] }
        briefUpdatedAt: { type: ["string", "null"] }
        avgDaysToPay: { type: ["number", "null"] }
        status: { type: string }
        createdAt: { type: ["string", "null"] }
    PropertySummary:
      type: object
      required: [id, name, owed, openJobs]
      properties:
        id: { type: string }
        name: { type: string }
        pmcName: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        units: { type: ["integer", "null"] }
        owed: { type: number }
        openJobs: { type: integer }
        marginPct: { type: ["number", "null"] }
    PropertyDetail:
      type: object
      required: [property, contacts, priceItems, jobs, expenses, agreements, stats]
      properties:
        property:
          $ref: "#/components/schemas/Property"
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contact"
        priceItems:
          type: array
          items:
            $ref: "#/components/schemas/PriceItem"
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        expenses:
          type: array
          items:
            $ref: "#/components/schemas/Expense"
        agreements:
          type: array
          items:
            $ref: "#/components/schemas/Agreement"
        stats:
          $ref: "#/components/schemas/PropertyStats"
    PropertyStats:
      type: object
      required: [owed, openJobs, marginPct, mtdRevenue]
      properties:
        owed: { type: number }
        openJobs: { type: integer }
        marginPct: { type: ["number", "null"] }
        mtdRevenue: { type: number }
    PropertyInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        pmcName: { type: string }
        city: { type: string }
        units: { type: integer }
        accessNotes: { type: string }
    PropertyUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        pmcName: { type: string }
        city: { type: string }
        units: { type: integer }
        accessNotes: { type: string }
        brief: { type: string }
        status: { type: string }
    Contact:
      type: object
      required: [id, name]
      properties:
        id: { type: string }
        propertyId: { type: ["string", "null"] }
        name: { type: string }
        role: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        prefers: { type: ["string", "null"] }
    ContactInput:
      type: object
      required: [name]
      properties:
        propertyId: { type: string }
        name: { type: string, minLength: 1 }
        role: { type: string }
        phone: { type: string }
        email: { type: string }
        prefers: { type: string }
    PriceItem:
      type: object
      required: [id, propertyId, service, rate]
      properties:
        id: { type: string }
        propertyId: { type: string }
        service: { type: string }
        detail: { type: ["string", "null"] }
        unit: { type: ["string", "null"] }
        rate: { type: number }
        marginFloor: { type: ["number", "null"] }
    PriceItemInput:
      type: object
      required: [service, rate]
      properties:
        service: { type: string, minLength: 1 }
        detail: { type: string }
        unit: { type: string }
        rate: { type: number }
        marginFloor: { type: number }
    Agreement:
      type: object
      required: [id, title]
      properties:
        id: { type: string }
        propertyId: { type: ["string", "null"] }
        title: { type: string }
        storagePath: { type: ["string", "null"] }
        effectiveFrom: { type: ["string", "null"] }
        renewsOn: { type: ["string", "null"] }
...
    Job:
      type: object
      required: [id, jobNo, status]
      properties:
        id: { type: string }
        jobNo: { type: string }
        woNo: { type: ["string", "null"] }
        propertyId: { type: string }
        propertyName: { type: ["string", "null"] }
        unitNo: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        description: { type: ["string", "null"] }
        status: { type: string }
        crewLeaderId: { type: ["string", "null"] }
        crewLeaderName: { type: ["string", "null"] }
        bidId: { type: ["string", "null"] }
        contactId: { type: ["string", "null"] }
        inspectionRequired: { type: ["boolean", "null"] }
        inspectionPassedAt: { type: ["string", "null"] }
        completedAt: { type: ["string", "null"] }
        recapSentAt: { type: ["string", "null"] }
        warrantyUntil: { type: ["string", "null"] }
        scheduledOn: { type: ["string", "null"] }
        grossProfit: { type: ["number", "null"] }
        marginPct: { type: ["number", "null"] }
        createdAt: { type: ["string", "null"] }
...
    Expense:
      type: object
      required: [id, amount]
      properties:
        id: { type: string }
        jobId: { type: ["string", "null"] }
        propertyId: { type: ["string", "null"] }
        vendor: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        amount: { type: number }
        source: { type: ["string", "null"] }
        spentOn: { type: ["string", "null"] }
...