# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * HALO — Archangel Operations Layer API
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: string;
  name: string;
  /** @nullable */
  pmcName?: string | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  units?: number | null;
  /** @nullable */
  accessNotes?: string | null;
  /** @nullable */
  brief?: string | null;
  /** @nullable */
  briefUpdatedAt?: string | null;
  /** @nullable */
  avgDaysToPay?: number | null;
  status: string;
  /** @nullable */
  createdAt?: string | null;
}