# lib/api-zod/src/generated/types/propertyUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * HALO — Archangel Operations Layer API
 * OpenAPI spec version: 0.1.0
 */

export interface PropertyUpdate {
  /** @minLength 1 */
  name?: string;
  pmcName?: string;
  city?: string;
  units?: number;
  accessNotes?: string;
  brief?: string;
  status?: string;
}